use ratio.py to generate nice html

+++ src/Classes/arrondissement.py
@@ -5,14 +5,14 @@ Represents a Borough
     def __init__(self, name, stops, area, population):
         self.name = name
         self.stops = stops
-        self.area = area
-        self.population = population
+        self.area = float(area)
+        self.population = int(population)
 
     def __repr__(self):
         return self.name + " has population " + str(self.population) + " and area " + str(self.area) + "km^-2"
 
     def density(self):
-        return self.population / self.area
+        return int(self.population / self.area)
 
     def ratio(self):
-        return self.stops / (float(self.population) * float(self.area))
+        return self.stops / (self.population * self.area)
